BaseEntity applied

+++ tempus-repository/src/main/java/in/batur/tempus/repository/CompanyRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import in.batur.tempus.entity.Company;
 
 @Repository
-public interface CompanyRepository extends PagingAndSortingRepository<Company, Integer>
+public interface CompanyRepository extends PagingAndSortingRepository<Company, Long>
 {
 
 }
