+++ tempus-repository/src/main/java/in/batur/tempus/repository/RoleRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import in.batur.tempus.entity.Role;
 
 @Repository
-public interface RoleRepository extends PagingAndSortingRepository<Role, Integer>
+public interface RoleRepository extends PagingAndSortingRepository<Role, Long>
 {
 
 }
